test(error_log): move to_jsonl wire-shape tests from pipeline
The doc on `ErroredEventContext::to_jsonl` already asserts that the
JSONL shape is `error_log`'s to own — `pipeline` builds the context
but the wire format (schema_version, forbidden routing fields on
Output, Event::from_json round-trip) is what the DLQ writer and any
downstream tooling read against. The five pure wire-shape tests were
still living under `pipeline::tests` because they had been there
before to_jsonl was moved. Relocate them next to the writer they
constrain, leaving a one-line pointer in pipeline's test module so a
reader looking for the shape assertions lands in the right place.
The unrelated pipeline-flow tests (`process_variant_named_process_site_selection`
etc.) stay put.
